backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
django-filter
python-dotenv
gunicorn
